+++ gantry/config.py
@@ -1,15 +1,19 @@
 import dataclasses
 from dataclasses import dataclass
-from typing import Literal, Optional
+from typing import Any, Literal, Optional
 
 import tomli
 from dataclass_extensions import decode
+
+from .exceptions import ConfigurationError
 
 
 @dataclass
 class GantryConfig:
     workspace: Optional[str] = None
     budget: Optional[str] = None
+    log_level: Optional[Literal["debug", "info", "warning", "error"]] = None
+    quiet: Optional[bool] = None
     path: Optional[str] = dataclasses.field(repr=False, default=None)
 
     @classmethod
@@ -25,15 +29,31 @@ class GantryConfig:
         except FileNotFoundError:
             return cls()
 
-        out = decode(cls, data.get("tool", {}).get("gantry", {}))
+        try:
+            out = decode(cls, data.get("tool", {}).get("gantry", {}))
+        except Exception as e:
+            raise ConfigurationError(
+                f"Failed to decode gantry config from {path} ({type(e).__name__}: {e})"
+            ) from e
+
         out.path = path
         return out
 
-    def get_help_string_for_default(self, field: Literal["workspace", "budget"]) -> str:
-        value = getattr(self, field)
-        if value is None:
-            return ""
-        elif self.path is not None:
-            return f"Defaults to '{value}' (from config in {self.path})."
+    def _get_value_string(self, value: Any) -> str:
+        if isinstance(value, bool):
+            return str(value).lower()
         else:
-            return f"Defaults to '{value}' (from config)."
+            return f"'{value}'"
+
+    def get_help_string_for_default(
+        self, field: Literal["workspace", "budget", "log_level", "quiet"], default: Any = None
+    ) -> str:
+        value = getattr(self, field)
+        if value is None and default is None:
+            return ""
+        elif value is None:
+            return f"Defaults to {self._get_value_string(default)}."
+        elif self.path is not None:
+            return f"Defaults to {self._get_value_string(value)} (from config in {self.path})."
+        else:
+            return f"Defaults to {self._get_value_string(value)} (from config)."