+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for setting certain gantry defaults in a `pyproject.toml` file under the `[tool.gantry]` section.
-  For now the only options are `workspace` and `budget`.
+  For now the only options are `workspace`, `budget`, `log_level`, and `quiet`.
 
 ## [v3.0.0](https://github.com/allenai/beaker-gantry/releases/tag/v3.0.0) - 2025-08-06