+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v3.1.0](https://github.com/allenai/beaker-gantry/releases/tag/v3.1.0) - 2025-08-25
+
 ### Added
 
 - Added support for setting certain gantry defaults in a `pyproject.toml` file under the `[tool.gantry]` section.